requirements.txt
google-generativeai
python-dotenv
wikipedia
pypdf